MyStoreProject/src/test/java/com/myStore/testScripts/AddToCartPageTest.java
package com.myStore.testScripts;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Parameters;
import org.testng.annotations.Test;

import com.myStore.base.BaseClass;
import com.myStore.pageObjects.AddToCartPage;
import com.myStore.pageObjects.HomePage;
import com.myStore.pageObjects.IndexPage;
import com.myStore.pageObjects.SearchResultPage;

public class AddToCartPageTest extends BaseClass {

	IndexPage indexPage;
	SearchResultPage loginPage;
	HomePage homePage;
	SearchResultPage searchResultPage;
	AddToCartPage addToCartPage;
	
	@Parameters("browser")
	@BeforeMethod(groups = {"Smoke","Sanity","Regression"})
	public void setup(String browser) {
		lunchApp(browser);
	}
	
	@AfterMethod(groups = {"Smoke","Sanity","Regression"})
	public void tearDown() {
		if (driver !=null) {
			driver.quit();
		}
	}
	
	@Test(groups = {"Regression","Sanity"})
	public void addToCartTest() {
		indexPage = new IndexPage();
		searchResultPage = indexPage.searchProduct("t-shirt");
		addToCartPage = searchResultPage.clickOnProduct();
		addToCartPage.sizeSelect("L");
		addToCartPage.enterQuantity("2");
		addToCartPage.clickOnAddToCart();
		boolean vlidateProductAddeddIntoAddToCart = addToCartPage.vlidateProductAddeddIntoAddToCart();
		Assert.assertTrue(vlidateProductAddeddIntoAddToCart);
		
	}
}
